More refactoring and dep upgrades
